add calcul mean

+++ src/main/java/database/Anime.java
@@ -80,7 +80,7 @@ public class Anime {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(double rating) {
         this.rating = rating;
     }
